up, down으로 통합

+++ start.bat
@@ -7,23 +7,17 @@ If  %1 == abnormaldown goto:abnormaldown
 
 :down
   echo "down docker containers of log anticheat..."
-  docker-compose -f kafka-docker-compose/docker-compose.yml -f elastic-kibana/docker-compose.yaml -f fluentd/docker-compose.yaml down --volume
-  docker-compose -f python3.7/docker-compose.yaml down --volume
+  If  %2 == kfek docker-compose -f kafka-docker-compose/docker-compose.yml -f elastic-kibana/docker-compose.yaml -f fluentd/docker-compose.yaml down --volume
+  If  %2 == log docker-compose -f python3.7/docker-compose.yaml down --volume
+  If  %2 == abnormal docker-compose -f abnormal-log-generator/docker-compose.yaml up -d --build
+
 goto:eof
 
 
 :up
   echo "start log anticheat..."
-  docker-compose -f kafka-docker-compose/docker-compose.yml -f elastic-kibana/docker-compose.yaml -f fluentd/docker-compose.yaml up -d --build
-  docker-compose -f python3.7/docker-compose.yaml up -d --build
-goto:eof
+  If  %2 == kfek docker-compose -f kafka-docker-compose/docker-compose.yml -f elastic-kibana/docker-compose.yaml -f fluentd/docker-compose.yaml up -d --build
+  If  %2 == log docker-compose -f python3.7/docker-compose.yaml up -d --build
+  If  %2 == abnormal docker-compose -f abnormal-log-generator/docker-compose.yaml down --volume
 
-:abnormalup
-  echo "start abnormal-log-generator..."
-  docker-compose -f abnormal-log-generator/docker-compose.yaml up -d --build
-goto:eof
-
-:abnormaldown
-  echo "down abnormal-log-generator..."
-  docker-compose -f abnormal-log-generator/docker-compose.yaml down --volume
 goto:eof
